move NetworkGeneralService to network_/general.py
